Fix JRUBY-6949
system() silently fails when options are passed

I believe we were trying to coerce the hash args to String, which
failed...and then the failure was swallowed due to another bug
(already filed). This commit does not add options support, but
does trim them off the end of the args array and warns (once) that
they will be ignored. We plan to get the remaining process options
working in 1.7.1, and there are numerous other bugs for that.