Make GC::Profiler.report stream the report.
This saves memory and potentially GC overhead by avoiding the
creation of a large string containing all the GC report lines. It
alters result by providing a way to pass in an "out" target that's
a blank String by default. Report now passes in $stdout as the
target of the report, so only the individual lines must be created
rather than the entire report String.

See also https://bugs.ruby-lang.org/issues/6865